backend/requirements.txt
flask==2.3.3
boto3
mysql-connector-python
serverless-wsgi

